Generate predefined puzzles up to 10.
I actually generated 100 puzzles in total, but when I tried to put them
all into the code, the VM Emulator crashed. I guess that's too much for
the little Hack computer to handle. So I commented out the bottom 90
puzzles but left them in there in case I decide to use them later.

It might be interesting to come up with a more compact binary
representation for puzzles, but since Jack doesn't have bit shifting
operations that would be a real pain to unpack.